refactor: Move Bootstrap skin to top level
